changed protobuf to a struct; refactored code to make it more modular
