+++ src-tauri/src/aux_functions/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod datatypes;
 pub(crate) mod edge_factory;
-pub(crate) mod neighbor_factory;
+pub(crate) mod graph_init;
